% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_time.R
\name{append_time}
\alias{append_time}
\title{Add suffix of the form \verb{_\%Y_\%m_\%d_\%H_\%M_\%S} to file name, keeping extension..}
\usage{
append_time(filename)
}
\arguments{
\item{filename}{File name as string.}
}
\value{
New filename as string.
}
\description{
Add suffix of the form \verb{_\%Y_\%m_\%d_\%H_\%M_\%S} to file name, keeping extension..
}
\author{
Fernando Naufel
}
