% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_final.R
\name{create_final}
\alias{create_final}
\title{Bind the 2 most recent reviewed dfs (matched and unmatched) in dirname.}
\usage{
create_final(dirname)
}
\arguments{
\item{dirname}{Directory where csv files are.}
}
\value{
A tibble.
}
\description{
This function saves the merged df as df_merged_9999_99_99_99_99_99.csv and also returns the df.
}
\author{
fnaufel
}
