% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_cat_and_city.R
\name{extract_cat_and_city}
\alias{extract_cat_and_city}
\title{Extracts category and city from \code{info} field.}
\usage{
extract_cat_and_city(df)
}
\arguments{
\item{df}{A tibble created by \code{\link[=extract_monthly_payments]{extract_monthly_payments()}}.}
}
\value{
A tibble.
}
\description{
Creates columns \code{categoria} and \code{cidade}.
}
\author{
Fernando Naufel
}
