% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_monthly_payments.R
\name{extract_monthly_payments}
\alias{extract_monthly_payments}
\title{Extract installment info from \code{item} field.}
\usage{
extract_monthly_payments(df)
}
\arguments{
\item{df}{A tibble created by \code{\link[=join_lines]{join_lines()}}.}
}
\value{
A tibble.
}
\description{
Creates columns \code{parcela} and \code{n_parcelas}.
}
\author{
Fernando Naufel
}
