% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_categories.R
\name{fix_categories}
\alias{fix_categories}
\title{Match items to more detailed categories and subcategories.}
\usage{
fix_categories(df, dirname)
}
\arguments{
\item{df}{Tibble produced by \code{\link[=guess_year]{guess_year()}}.}

\item{dirname}{Path passed by the user to \code{\link[=import_dir]{import_dir()}}.}
}
\value{
A tibble.
}
\description{
This function uses file \code{../../items_categories.csv}, which contains the association between items and categories.
}
\author{
Fernando Naufel
}
