% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_year.R
\name{guess_year}
\alias{guess_year}
\title{Compute the year of each item's date.}
\usage{
guess_year(df, due_date)
}
\arguments{
\item{df}{Tibble produced by \code{\link[=extract_cat_and_city]{extract_cat_and_city()}}.}

\item{due_date}{String containing date of current fatura in format 'dd/mm/yyyy'.}
}
\value{
A tibble.
}
\description{
Originally, \code{data} only has day and month. We compare \code{data} with the due date (usually this month); if \code{data} is in the future, we conclude the item is an installment from last year.

\strong{Careful:} this only works \emph{if no purchase has more than 12 installments}.
}
\author{
Fernando Naufel
}
