% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_csv.R
\name{import_csv}
\alias{import_csv}
\title{Import a csv file produced by \href{https://tabula.technology/}{tabula}.}
\usage{
import_csv(filename)
}
\arguments{
\item{filename}{Path and name of the csv file.}
}
\value{
A tibble.
}
\description{
The user must manually select the areas of the pdf file to be converted to csv files.

The csv file must not have column names in the first line.

The csv file must contain 3 columns, otherwise, an error is thrown.
}
\author{
Fernando Naufel
}
