% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_lines.R
\name{join_lines}
\alias{join_lines}
\title{Gathers info for each payment.}
\usage{
join_lines(df)
}
\arguments{
\item{df}{A tibble produced by \code{\link[=import_dir]{import_dir()}}.}
}
\value{
A tibble.
}
\description{
Because of the way the pdf file is converted to csv, each record spans 2 lines. The second line contains info about the category and the city.
}
\author{
Fernando Naufel
}
