% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_df.R
\name{read_df}
\alias{read_df}
\title{Read tibble from csv file.}
\usage{
read_df(filename)
}
\arguments{
\item{filename}{Csv file to be read, with path.}
}
\value{
A tibble.
}
\description{
We expect the csv file is in the format generated by \code{\link[DataEditR:data_edit]{DataEditR::data_edit()}}:
\itemize{
\item First column is row numbers.
\item Date columns are chr.
\item Valor is chr because of comma decimal mark; must be read with locale.
}
}
\author{
fnaufel
}
