% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review.R
\name{review}
\alias{review}
\title{Read a csv file as produced by \code{\link[=import_dir]{import_dir()}} into a tibble and edit interactively.}
\usage{
review(filename, categories = "../../categories.csv")
}
\arguments{
\item{filename}{Path and name of csv file.}

\item{categories}{Path and name of csv file containing categories.}
}
\description{
This function uses a table of categories for autocompletion.
}
