% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review.R
\name{review_unmatched}
\alias{review_unmatched}
\title{Read \code{df_unmatched.csv} or \code{df_unmatched_reviewed.csv} into a tibble and edit interactively.}
\usage{
review_unmatched(dirname)
}
\arguments{
\item{dirname}{Directory where csv files are.}
}
\description{
If the \verb{_reviewed} file exists, edit that. It may have partial changes made by the user.
}
