% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_history.R
\name{use_history}
\alias{use_history}
\title{Try to get info from last month's csv to fix categories.}
\usage{
use_history(df, last_month_final_csv)
}
\arguments{
\item{df}{This month's tibble.}

\item{last_month_final_csv}{Path to last month's csv file.}
}
\value{
A tibble.
}
\description{
Try to get info from last month's csv to fix categories.
}
\author{
Fernando Naufel
}
