% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_df.R
\name{write_df}
\alias{write_df}
\title{Write tibble to csv file.}
\usage{
write_df(df, filename)
}
\arguments{
\item{df}{Tibble to write.}

\item{filename}{Csv file to write, with path.}
}
\value{
A tibble.
}
\description{
The csv file is in the format generated by \code{\link[DataEditR:data_edit]{DataEditR::data_edit()}}:
\itemize{
\item First column is row numbers.
\item Date columns are chr.
\item Valor is chr because of comma decimal mark; must be read with locale.
}
}
\author{
fnaufel
}
