% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_final.R
\name{create_final}
\alias{create_final}
\title{Merge the two reviewed tibbles, save and return the result.}
\usage{
create_final(dirname)
}
\arguments{
\item{dirname}{Directory where csv files are.}
}
\value{
A tibble.
}
\description{
Merge the two reviewed tibbles, save and return the result.
}
\author{
fnaufel
}
