% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_dir.R
\name{import_dir}
\alias{import_dir}
\title{Imports all csv files in directory \code{dirname}.}
\usage{
import_dir(dirname)
}
\arguments{
\item{dirname}{Directory as string.}
}
\value{
A tibble.
}
\description{
Just calls \code{\link[=import_csv]{import_csv()}} on all files with extension \code{.csv}.
}
\author{
Fernando Naufel
}
