% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_dir.R
\name{import_dir}
\alias{import_dir}
\title{Import and process all tabula csv files in directory \code{dirname}.}
\usage{
import_dir(dirname, due_date)
}
\arguments{
\item{dirname}{Directory as string.}

\item{due_date}{Date of current fatura.}
}
\description{
This will read all csv files of the form 'tabula*.csv' in \code{dirname} and create 2 files (also in \code{dirname}):
\itemize{
\item \code{df_matched.csv}, containing records whose categories were automatically determined (based on a table read from \code{dirname/../../categories.csv}), and
\item \code{df_unmatched.csv}, containing records with the original categories from the fatura pdf.
}
}
\author{
Fernando Naufel
}
