% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/review.R
\name{review_matched}
\alias{review_matched}
\title{Read \code{df_matched.csv} into a tibble and edit interactively.}
\usage{
review_matched(dirname)
}
\arguments{
\item{dirname}{Directory where csv files are.}
}
\description{
Read \code{df_matched.csv} into a tibble and edit interactively.
}
